aws-cdk-cis/requirements.txt
-e .
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-logs
aws-cdk.aws-guardduty
aws-cdk.aws-cloudtrail
aws-cdk.aws-s3
aws-cdk.aws-config
aws-cdk.aws-kms
pylint
pep8
